tests/wpt/tests/css/css-cascade/scope-import-scope-start.tentative.html

<!DOCTYPE html>
<title>@import scope(), &lt;scope-start&gt;</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/7348">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
@import url("resources/scope-imported.css") scope((.scope));
</style>
<main id=main>
  <div class=scope>
    <div class=x>Inside</div>
  </div>
  <div class=x>Outside</div>
</main>
<script>
  test(() => {
    let e = main.querySelector('.scope > .x');
    assert_equals(getComputedStyle(e).getPropertyValue('--x'), '1');
  }, 'Scope-imported rule applies within scope');

  test(() => {
    let e = main.querySelector('main > .x');
    assert_equals(getComputedStyle(e).getPropertyValue('--x'), '');
  }, 'Scope-imported rule does not apply outside of scope');
</script>
